Disallow matching on all-private structs.
This crate was relying on changing all-private structs to enums being
semver-compatible even though it's not actually the case because of an
edge case with matching on such structs being allowed. This updates the
documentation to disallow such matching preserving future proofing even
though it can't be enforced by the compiler (yet).

See also https://internals.rust-lang.org/t/disallow-matching-on-all-private-structs/15993