Removed unused features
After better analysis it was determined that these features don't help
with forward compatibility and only slightly improve convenience. The
noise is probably not worth it.

`bitcoin_*` were somewhat more interesting because the `bitcoin` crate
has sad `no_std` support but there's hope it'll be improved and even if
not adding features is better than removing. (The features weren't
released yet.)

Closes #10